+++ AddOns/Details/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Details! Damage Meter
 
-## [Details.13132.161](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.13132.161) (2024-10-18)
-[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.13131.161...Details.13132.161) 
+## [Details.13166.161](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.13166.161) (2024-11-05)
+[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.13132.161...Details.13166.161) 
 
-- Updating framework  
+- Framework and bug fixes  
+- Fixed an issue with classic versions of the game  
+- fixing the latest commit  
+- Cache challenge mode mapID at the start of the run  
+- Using xpcall for events instead of pcall  
+- Updated GetChallengeCompletionInfo  
+- Fixed GetCompletionInfo  
+- Updates